[AArch64] Stricter ldp/stp alignment guaranteed
When ldp-aligned-only/stp-aligned-only is specified, modified to
cancel ldp/stp transformation if MachineMemOperand is not present or
the access size is unknown.
In the previous implementation, the test passed when there was no
MachineMemOperand. Also, if the size was unknown, an incorrect value
was used or an assertion failed. (But actually, if there is no
MachineMemOperand, it will be excluded from the target by
isCandidateToMergeOrPair() before reaching the part.)

A statistic NumFailedAlignmentCheck is added. NumPairCreated is
modified so that it only counts if it is not cancelled.